requirements.txt
torch
torchvision
pandas
numpy
matplotlib
scikit-learn
PyYAML
openpyxl
fastapi
uvicorn
pillow
